requirement.txt
streamlit
tensorflow
Pillow
numpy
rembg
onnxruntime
pandas
streamlit-cropper